Fixed the bugs
Fixed where you could make a blank user and password. Fixed Where you could go to the sign on or create account even if you were logged in. Fixed table issue where it was showing the teacher id. api now pulls a whole teacher object instead of just an id. fixed the link on the "Pidgeon Final Project" so it goes to index instead of sign up. If you are not signed in the index page or any other page will send you to the sign on page.